give deflate window its own file
